Move some types around to base for consistency with patterns used elsewhere.
